% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.opt.R
\name{A.opt}
\alias{A.opt}
\title{The A-optimality function}
\usage{
A.opt(K)
}
\arguments{
\item{K}{A numeric matrix of kinship.}
}
\value{
This function will return each individuals' A-optimality values.
}
\description{
This function is designed to calculate A-optimality.
}
\examples{
data(kinship)
A.opt(kinship)
}
