% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD.R
\name{CD}
\alias{CD}
\title{The CD function}
\usage{
CD(K)
}
\arguments{
\item{K}{A numeric matrix of kinship.}
}
\value{
This function will return each individuals' CD values.
}
\description{
This function is designed to calculate CD.
}
\examples{
data(kinship)
CD(kinship)
}
