% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDmeanv2.R
\name{CDmeanv2}
\alias{CDmeanv2}
\title{The CDmean(v2) function}
\usage{
CDmeanv2(K, train, test)
}
\arguments{
\item{K}{A numeric matrix of kinship.}

\item{train}{An integer vector of which individuals are in the training set.}

\item{test}{An integer vector of which individuals are in the test set.}
}
\value{
This function will return the CDmean(v2) value.
}
\description{
This function is designed to calculate CDmean(v2).
}
\examples{
data(kinship)
CDmeanv2(kinship,1:50,1:100)
}
