% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDranking.R
\name{CDranking}
\alias{CDranking}
\title{The CDranking function}
\usage{
CDranking(K)
}
\arguments{
\item{K}{A numeric matrix of kinship.}
}
\value{
This function will return each individuals' CDranking values.
}
\description{
This function is designed to calculate CDranking.
}
\examples{
data(kinship)
CDranking(kinship)
}
