% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D.opt.R
\name{D.opt}
\alias{D.opt}
\title{The D-optimality function}
\usage{
D.opt(K)
}
\arguments{
\item{K}{A numeric matrix of kinship.}
}
\value{
This function will return D-optimality value.
}
\description{
This function is designed to calculate D-optimality.
}
\examples{
data(PC)
D.opt(kinship[1:50,1:50])
}
