% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA.R
\name{GA}
\alias{GA}
\title{The Genetic Algorithm function}
\usage{
GA(
  K = NULL,
  geno = NULL,
  cand,
  n.train,
  subpop = NULL,
  test = NULL,
  criteria = r_score,
  greater = TRUE,
  min.iter = NULL,
  max.iter = NULL,
  console = TRUE
)
}
\arguments{
\item{K}{A numeric matrix of kinship. (for GVoverall and CDmean(v2) method)}

\item{geno}{A numeric matrix of marker-associated matrix or principal components (for rSocre and MSPE methods; rows: individuals; columns: PCs or markers).}

\item{cand}{An integer vector of which individuals are candidates of the training set.}

\item{n.train}{The size of the training set.}

\item{subpop}{A character vector of subpopulation.}

\item{test}{An integer vector of which individuals are in the test set.}

\item{criteria}{Select the criteria uses to discover the best genotypes (deafult: r_score).}

\item{greater}{A boolean variable determines whether the criteria is the greater, the better.}

\item{min.iter}{Minimum iteration.}

\item{max.iter}{Maximum iteration.}

\item{console}{Whether the function will print out the iteration numbers (deafult: TRUE).}
}
\value{
This function will return 3 information including OPTtrain (a vector of chosen optimal training set), TOPscore (highest scores of before iteration), and ITERscore (criteria scores of each iteration).
}
\description{
This function is designed to perform genetic algorithm modified form Ou and Liao (2019).
}
\examples{
data(PC)
GA(geno=PC[,1:50], cand = 1:328, n.train = 50, min.iter = 50)

}
