% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVaverage.R
\name{GVaverage}
\alias{GVaverage}
\title{The GV-average function}
\usage{
GVaverage(K)
}
\arguments{
\item{K}{A numeric matrix of kinship.}
}
\value{
This function will return each individuals' GV-average values.
}
\description{
This function is designed to calculate GV-average.
}
\examples{
data(kinship)
GVaverage(kinship)
}
