% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVoverall.R
\name{GVoverall}
\alias{GVoverall}
\title{The GV-overall function}
\usage{
GVoverall(K)
}
\arguments{
\item{K}{A numeric matrix of kinship.}
}
\value{
This function will return GV-overall value.
}
\description{
This function is designed to calculate GV-overall.
}
\examples{
data(PC)
GVoverall(kinship[1:50,1:50])
}
