% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCD.R
\name{MCD}
\alias{MCD}
\title{The MCD function}
\usage{
MCD(K)
}
\arguments{
\item{K}{A numeric matrix of kinship.}
}
\value{
This function will return each individuals' MCD values.
}
\description{
This function is designed to calculate MCD.
}
\examples{
data(kinship)
MCD(kinship)
}
