% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspe.R
\name{mspe_score}
\alias{mspe_score}
\title{The MSPE score function}
\usage{
mspe_score(X, X0, lambda = 1)
}
\arguments{
\item{X}{A numeric matrix of the marker-associated matrices for the training set.}

\item{X0}{A numeric matrix of the marker-associated matrices for the candidate population.}

\item{lambda}{An integer number of the shrinkage parameter (deafult: 1)}
}
\value{
This function will return the MSPE score value.
}
\description{
This function is designed to calculate MSPE score.
}
\examples{
data(PC)
mspe_score(PC[1:50,],PC[51:100,])
}
