% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PC}
\alias{PC}
\title{Principal components matrix}
\format{
A numeric matrix with 328 rows and 328 columns (rows: individuals; columns: PCs).
}
\usage{
PC
}
\description{
Principal components matrix of Tropcial rice data set from Spindel et al. (2015).
}
\examples{
data(PC)
}
\references{
\url{doi.org/10.1371/journal.pgen.1004982}
}
\keyword{datasets}
