% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscore.R
\name{r_score}
\alias{r_score}
\title{The rScore function}
\usage{
r_score(X, X0)
}
\arguments{
\item{X}{A numeric matrix of the marker-associated matrices for the training set.}

\item{X0}{A numeric matrix of the marker-associated matrices for the candidate population.}
}
\value{
This function will return the rScore value.
}
\description{
This function is designed to calculate rScore from Ou and Liao (2019).
}
\examples{
data(PC)
r_score(PC[1:50,],PC[51:100,])
}
