% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSODBG.R
\name{TSODBG}
\alias{TSODBG}
\title{Training Set Optimization to Discover the Best Genotypes}
\usage{
TSODBG(
  K = NULL,
  geno = NULL,
  cand,
  n.train,
  subpop = NULL,
  test = NULL,
  method = "rScore",
  min.iter = NULL,
  max.iter = NULL,
  console = TRUE
)
}
\arguments{
\item{K}{A numeric matrix of kinship. (for  GVoverall, GVaverage, CDmean(v2), and CDranking methods)}

\item{geno}{A numeric matrix of marker-associated matrix or principal components (for rSocre and MSPE methods; rows: individuals; columns: PCs or markers).}

\item{cand}{An integer vector of which individuals are candidates of the training set.}

\item{n.train}{The size of the training set.}

\item{subpop}{A character vector of subpopulation.}

\item{test}{An integer vector of which individuals are in the test set.}

\item{method}{Select the method uses to discover the best genotypes (deafult: rScore).}

\item{min.iter}{Minimum iteration.}

\item{max.iter}{Maximum iteration.}

\item{console}{Whether the function will print out the iteration numbers (deafult: TRUE).}
}
\value{
This function will return the best genotypes.
}
\description{
This function is designed for discover the best genotypes.
}
\examples{
data(PC)
TSODBG(geno=PC[,1:50], cand = 1:328, n.train = 50, min.iter = 50)

}
